# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/iam/authentication/signIn": {
      "post": {
        "operationId": "AuthenticationController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User sign in data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationCredentialsModel"
                }
              }
            }
          }
        },
        "summary": "Sign in a user and set authentication cookies",
        "tags": [
          "Authentication",
          "Authentication"
        ]
      }
    },
    "/iam/authentication/refresh": {
      "post": {
        "operationId": "AuthenticationController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Refresh session data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication cookies refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationCredentialsModel"
                }
              }
            }
          }
        },
        "summary": "Refresh authentication cookies",
        "tags": [
          "Authentication",
          "Authentication"
        ]
      }
    },
...
    "/iam/authorization/roles": {
      "post": {
        "description": "Creates a new role in the system with the provided information including permissions.",
        "operationId": "AuthorizationController_createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Role creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role has been successfully created."
          },
          "400": {
            "description": "Bad Request - Invalid input data."
          },
          "409": {
            "description": "Conflict - Role with this title already exists."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create a new role",
        "tags": [
          "Authorization Management"
        ]
      },
      "get": {
        "operationId": "AuthorizationController_listRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Roles have been successfully retrieved."
          },
          "400": {
            "description": "Bad Request - Invalid input data."
          },
          "401": {
            "description": "Unauthorized - Invalid credentials."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "List roles",
        "tags": [
          "Authorization Management"
        ]
      }
    },
    "/iam/authorization/roles/{id}": {
      "get": {
        "operationId": "AuthorizationController_getRoleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role has been successfully retrieved."
          },
          "404": {
            "description": "Role not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Get a role by id",
        "tags": [
          "Authorization Management"
        ]
      },
      "patch": {
        "operationId": "AuthorizationController_updateRoleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role has been successfully updated."
          },
          "404": {
            "description": "Role not found."
          },
          "409": {
            "description": "Conflict - Role title already exists."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Update a role by id",
        "tags": [
          "Authorization Management"
        ]
      },
      "delete": {
        "operationId": "AuthorizationController_deleteRoleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role has been successfully deleted."
          },
          "404": {
            "description": "Role not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Delete a role by id",
        "tags": [
          "Authorization Management"
        ]
      }
    },
    "/iam/authorization/permissions": {
      "get": {
        "operationId": "AuthorizationController_getPermissionGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Permissions have been successfully retrieved."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Get all permission groups",
        "tags": [
          "Authorization Management"
        ]
      }
    },
    "/iam/authorization/permissions/{permissionGroup}": {
      "get": {
        "operationId": "AuthorizationController_getPermissions",
        "parameters": [
          {
            "name": "permissionGroup",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permissions have been successfully retrieved."
          },
          "404": {
            "description": "Permission group not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Get permissions by group",
        "tags": [
          "Authorization Management"
        ]
      }
    }
  },
  "info": {
    "title": "Support Service API",
    "description": "The Support Service API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SignInDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Identifier of the user",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password for authentication",
            "example": "SecurePassword123!"
          },
          "platform": {
            "type": "string",
            "description": "Platform of the client",
            "example": "web",
            "enum": [
              "web",
              "mobile",
              "desktop"
            ]
          }
        },
        "required": [
          "identifier",
          "password",
          "platform"
        ]
      },
      "AuthenticationCredentialsModel": {
        "type": "object",
        "properties": {}
      },
      "RefreshSessionDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "The access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "The refresh token"
          },
          "sessionId": {
            "type": "string",
            "description": "The session ID"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "sessionId"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the role",
            "example": "Admin",
            "minLength": 1,
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "description": "Color code for the role",
            "example": "#FF5733"
          },
          "description": {
            "type": "string",
            "description": "Description of the role",
            "example": "Full system administrator with all permissions",
            "maxLength": 255
          },
          "includedPermissions": {
            "type": "array",
            "description": "List of permissions included in this role",
            "example": [
              "CREATE_ROLE",
              "UPDATE_ROLE",
              "GET_ROLE",
              "LIST_ROLES"
            ],
            "items": {
              "type": "string",
              "enum": [
                "CREATE_ROLE",
                "UPDATE_ROLE",
                "DELETE_ROLE",
                "GET_ROLE",
                "LIST_ROLES"
              ]
            }
          },
          "excludedPermissions": {
            "type": "array",
            "description": "List of permissions specifically excluded from this role",
            "example": [
              "DELETE_ROLE"
            ],
            "items": {
              "type": "string",
              "enum": [
                "CREATE_ROLE",
                "UPDATE_ROLE",
                "DELETE_ROLE",
                "GET_ROLE",
                "LIST_ROLES"
              ]
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the role",
            "example": "Senior Admin",
            "minLength": 1,
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "description": "Color code for the role",
            "example": "#FF5733"
          },
          "description": {
            "type": "string",
            "description": "Description of the role",
            "example": "Updated role description",
            "maxLength": 255
          },
          "includedPermissions": {
            "type": "array",
            "description": "List of permissions included in this role",
            "example": [
              "CREATE_ROLE",
              "UPDATE_ROLE",
              "GET_ROLE"
            ],
            "items": {
              "type": "string",
              "enum": [
                "CREATE_ROLE",
                "UPDATE_ROLE",
                "DELETE_ROLE",
                "GET_ROLE",
                "LIST_ROLES"
              ]
            }
          },
          "excludedPermissions": {
            "type": "array",
            "description": "List of permissions specifically excluded from this role",
            "example": [
              "DELETE_ROLE"
            ],
            "items": {
              "type": "string",
              "enum": [
                "CREATE_ROLE",
                "UPDATE_ROLE",
                "DELETE_ROLE",
                "GET_ROLE",
                "LIST_ROLES"
              ]
            }
          }
        }
      }
    }
  }
}